refactor: extracted mailbox members router
