+++ apps/nextjs/src/trpc/router/mailbox/index.ts
@@ -1,6 +1,5 @@
 import { currentUser } from "@clerk/nextjs/server";
 import { TRPCError, type TRPCRouterRecord } from "@trpc/server";
-import { subHours } from "date-fns";
 import { and, count, eq, isNotNull, isNull, SQL } from "drizzle-orm";
 import { z } from "zod";
 import { setupOrganizationForNewUser } from "@/auth/lib/authService";
@@ -11,12 +10,12 @@ import { inngest } from "@/inngest/client";
 import { getLatestEvents } from "@/lib/data/dashboardEvent";
 import { getMailboxInfo } from "@/lib/data/mailbox";
 import { getClerkOrganization } from "@/lib/data/organization";
-import { getMemberStats } from "@/lib/data/stats";
 import { protectedProcedure } from "@/trpc/trpc";
 import { conversationsRouter } from "./conversations/index";
 import { customersRouter } from "./customers";
 import { faqsRouter } from "./faqs";
 import { githubRouter } from "./github";
+import { membersRouter } from "./members";
 import { metadataEndpointRouter } from "./metadataEndpoint";
 import { mailboxProcedure } from "./procedure";
 import { slackRouter } from "./slack";
@@ -98,31 +97,14 @@ export const mailboxRouter = {
     .mutation(async ({ ctx, input }) => {
       await db.update(mailboxes).set(input).where(eq(mailboxes.id, ctx.mailbox.id));
     }),
-  members: mailboxProcedure
-    .input(
-      z.object({
-        period: z.enum(["24h", "7d", "30d", "1y"]),
-        customDate: z.date().optional(),
-      }),
-    )
-    .query(async ({ ctx, input }) => {
-      const now = new Date();
-      const periodInHours = {
-        "24h": 24,
-        "7d": 24 * 7,
-        "30d": 24 * 30,
-        "1y": 24 * 365,
-      } as const;
 
-      const startDate = input.customDate || subHours(now, periodInHours[input.period]);
-      return await getMemberStats(ctx.mailbox, { startDate, endDate: now });
-    }),
   latestEvents: mailboxProcedure
     .input(z.object({ cursor: z.date().optional() }))
     .query(({ ctx, input }) => getLatestEvents(ctx.mailbox, input.cursor)),
+
   conversations: conversationsRouter,
   faqs: faqsRouter,
-
+  members: membersRouter,
   slack: slackRouter,
   github: githubRouter,
   tools: toolsRouter,
