+++ apps/nextjs/src/trpc/router/mailbox/members.ts
@@ -1,0 +1,27 @@
+import { type TRPCRouterRecord } from "@trpc/server";
+import { subHours } from "date-fns";
+import { z } from "zod";
+import { getMemberStats } from "@/lib/data/stats";
+import { mailboxProcedure } from "./procedure";
+
+export const membersRouter = {
+  stats: mailboxProcedure
+    .input(
+      z.object({
+        period: z.enum(["24h", "7d", "30d", "1y"]),
+        customDate: z.date().optional(),
+      }),
+    )
+    .query(async ({ ctx, input }) => {
+      const now = new Date();
+      const periodInHours = {
+        "24h": 24,
+        "7d": 24 * 7,
+        "30d": 24 * 30,
+        "1y": 24 * 365,
+      } as const;
+
+      const startDate = input.customDate || subHours(now, periodInHours[input.period]);
+      return await getMemberStats(ctx.mailbox, { startDate, endDate: now });
+    }),
+} satisfies TRPCRouterRecord;
